T3.6 - Working on SPI1 and Asynch
Found the code was not working on the T3.6 (actually probably all T3.x) boards when on SPI1 as  then DC must use it and so usage of standard IO pin for CS, needed to make sure alll data was output... So needed those places in code that did writedata and it was the last output needed to call writedata_last.

Also fixed updateScreenAsync so DMA updates is now working.