ST7789: fix for 240x320 display and no CS displays
Needed to fix a few places in the code that was storing width and height
or the like
in an uint8_t variable, which was fine until we went up to display that
now has 320.

So the Adafruit new display now works.

Also there are a lot of ST7789 displays out there that do not have a CS
connection brought out.

You can find a lot of these up on Amazon or Ebay.

Turns out you can make these displays work if you change to use
SPI_MODE2.  So like the Adafruit display
I changed the ST7789_t3.init method to have an optional SPI mode
parameter that you can pass in, which
defaults to SPI_MODE0.   So changed all of the code to pre build a
SPISettings structure with the
proper speed and mode built in and have the init code use it.

Was running into issue with T4 with this, that it would not pick it up
in the init time so added a hack to
transfer a 0 byte after setting it initially, which worked.

Also the code was changed to allow you to pass in -1 for cs, and the
code checks for it and does not
try to operate on it.