Fix 240x320 DMA and support for mini 160x80
The new ST7789 display that is 240x320 had issues with DMA code as I expected to do in 2 linked dmasettings.  But with these the size would exceed 64k per dmasetting.  So setup to use third one in this case.

Also added support for Adafruit mini 160x80 display.  Still need to test this when display arrives.