T3.x - MOSI/SCK on SPI pins use hardware SPI
Before if DC pin was NOT a hardware CS pin, the whole thing downgraded to using Software bitbang output...

In this version we detect this and if the main  MOSI/SCK, then setup for the actual SPI buss.  So _pspi->beginTransaction works, likewise we use _pspi->transfer instead of bitbang, plus now added function to also use _pspi->transfer16 if we are trying to output 16 bit values, like colors.