T3.6 DMA updates did not work
Issue with DMASettings on T3.6 when they are part of the main class object.  The constructor does not appear to be called to properly initialize them.  So created static member array, which wastes some memory as I create enough for 3 instances (one per buss)

Tested with three displays connected running DMA updates.