spitftbitmap - Fix color - show more ST77xx options
Added in some of the different options to uncomment for the different displays that we support.

Also the order of colors was wrong.  Showed up correct on ST7735 128x128 when using the wrong initializer .  But that would show up all other things done with colors wrong.

So added a few fill screens of Red, Green and Blue at the beginning with the color name printed in black in center of display so you would hopefully see that.  There is a half second delay between colors shown.

Note: I also had Arduino format the program so it shows some other whitspace things changing as well