Update uncannyEyes_async_st7789_240x240.ino
Was Faulting on T4.1 - With memory with code that tries to guess how much Stack space was used up by the code.

Needed to start fill farther away from _ebss.

Also while at it put in a new #define at start which can turn off this code, which defaults to not being done.

Also changed the default to not show the debug information on the eyes.