Fix Text center/charbounds issue like ili... drivers
There is an issue that charBounds did not use same calculations as actually used in the drawing of the text.  In most cases did not matter but for example
output of "-" will not offset correctly and y will be near bottom

Also for centered text, issue where it dived the offset to the starting of the drawing in half where it should be the whole offset.