Update the St77xx T4 code to not copy to internal buffers
The DMA code for the T4.x was based on my earlier DMA code, to avoid DMA issues on T4.x where the Frame buffer was in DMAMEM or PS memory...
And the graphic primitives would complete and we start up a DMA operation,  The updated data may not go out do to the cache...  So I copied it to low memory which worked, but caused lots of memory copy and interrupts.

Later I switched to be more like other Teensy code, where it sets up 3 DMASettings objecjts chained to each other, and on single shot DMA output, take care of the issue with arm_dcache_... call at startup of DMA.  And for continuous, I have interrupt at end of each frame, that does it again.

I also added the ili9341_t3n code for allowing callbacks to code when a frame completes and optionally at half frame completing.

So far only tested on T4 branch (MicroMod)

Also added the callback code for full and half frames to the T3.5 and T3.6 code.
So far it has been tested on ST7789 240x240 and 320x240 on T3.5 and T3.6,
but only on SPI.