Fix some pin names and constructors
For the Verbose version, there was an ST7735 and an ST7789 version, but the ST7735 was accidentally edited to ST7789 version... So fixed that but still left the ST7789 as default one by uncommenting the appropriate lines.

Also fixed teh short version of the ST7735 constructor which was using (cs, dc, rst) instead of what all of the others were using (TFT_CS, TFT_DC, TFT_RST)

Two examples were this way.