Added the code to only update changed areas
Copied over the function calls from the ILI9341_t3n.  There is maybe some more functions missing the updates.

But also added in new code from ILI9341_t3n the code now that when you do writeRect it looks and only marks the dirty area for the bounding rectangle of changed pixels (which may be the whole thing